medicar/requirements.txt
python-decouple==3.4
Django==3.1.4
django-cors-headers==3.8.0
django-filter==2.4.0
djangorestframework==3.12.4
django-filter==2.4.0
pytz==2021.1
sqlparse==0.4.2
asgiref==3.4.1
Markdown==3.3.4
